Refactoring and force CM to check for system.yaml before including system.
